Add abstraction of parameters for JDBC connections
