+++ presto-jdbc/src/main/java/com/facebook/presto/jdbc/PrestoConnection.java
@@ -16,10 +16,7 @@ package com.facebook.presto.jdbc;
 import com.facebook.presto.client.ClientSession;
 import com.facebook.presto.client.ServerInfo;
 import com.facebook.presto.client.StatementClient;
-import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.net.HostAndPort;
 import io.airlift.units.Duration;
 
 import java.net.URI;
@@ -41,7 +38,6 @@ import java.sql.SQLXML;
 import java.sql.Savepoint;
 import java.sql.Statement;
 import java.sql.Struct;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Properties;
@@ -53,9 +49,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Strings.isNullOrEmpty;
 import static com.google.common.collect.Maps.fromProperties;
-import static io.airlift.http.client.HttpUriBuilder.uriBuilder;
 import static java.nio.charset.StandardCharsets.US_ASCII;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MINUTES;
@@ -64,31 +58,25 @@ public class PrestoConnection
         implements Connection
 {
     private final AtomicBoolean closed = new AtomicBoolean();
-    private final AtomicReference<String> catalog = new AtomicReference<>();
-    private final AtomicReference<String> schema = new AtomicReference<>();
     private final AtomicReference<String> timeZoneId = new AtomicReference<>();
     private final AtomicReference<Locale> locale = new AtomicReference<>();
-    private final URI uri;
-    private final HostAndPort address;
+
+    private final ConnectionParameters connectionParameters;
+
     private final String user;
     private final Map<String, String> clientInfo = new ConcurrentHashMap<>();
     private final Map<String, String> sessionProperties = new ConcurrentHashMap<>();
     private final AtomicReference<String> transactionId = new AtomicReference<>();
     private final QueryExecutor queryExecutor;
 
-    PrestoConnection(URI uri, String user, QueryExecutor queryExecutor)
+    PrestoConnection(ConnectionParameters connectionParameters, String user, QueryExecutor queryExecutor)
             throws SQLException
     {
-        this.uri = requireNonNull(uri, "uri is null");
-        this.address = HostAndPort.fromParts(uri.getHost(), uri.getPort());
+        this.connectionParameters = connectionParameters;
         this.user = requireNonNull(user, "user is null");
         this.queryExecutor = requireNonNull(queryExecutor, "queryExecutor is null");
         timeZoneId.set(TimeZone.getDefault().getID());
         locale.set(Locale.getDefault());
-
-        if (!isNullOrEmpty(uri.getPath())) {
-            setCatalogAndSchema();
-        }
     }
 
     @Override
@@ -203,7 +191,7 @@ public class PrestoConnection
             throws SQLException
     {
         checkOpen();
-        this.catalog.set(catalog);
+        connectionParameters.setCatalog(catalog);
     }
 
     @Override
@@ -211,7 +199,7 @@ public class PrestoConnection
             throws SQLException
     {
         checkOpen();
-        return catalog.get();
+        return connectionParameters.getCatalog();
     }
 
     @Override
@@ -472,7 +460,7 @@ public class PrestoConnection
             throws SQLException
     {
         checkOpen();
-        this.schema.set(schema);
+        connectionParameters.setSchema(schema);
     }
 
     @Override
@@ -480,7 +468,7 @@ public class PrestoConnection
             throws SQLException
     {
         checkOpen();
-        return schema.get();
+        return connectionParameters.getSchema();
     }
 
     public String getTimeZoneId()
@@ -562,7 +550,7 @@ public class PrestoConnection
 
     URI getURI()
     {
-        return uri;
+        return connectionParameters.getURI();
     }
 
     String getUser()
@@ -572,12 +560,12 @@ public class PrestoConnection
 
     ServerInfo getServerInfo()
     {
-        return queryExecutor.getServerInfo(getHttpUri());
+        return queryExecutor.getServerInfo(connectionParameters.getHttpUri());
     }
 
     StatementClient startQuery(String sql)
     {
-        URI uri = getHttpUri();
+        URI uri = connectionParameters.getHttpUri();
 
         String source = firstNonNull(clientInfo.get("ApplicationName"), "presto-jdbc");
 
@@ -585,8 +573,8 @@ public class PrestoConnection
                 uri,
                 user,
                 source,
-                catalog.get(),
-                schema.get(),
+                connectionParameters.getCatalog(),
+                connectionParameters.getSchema(),
                 timeZoneId.get(),
                 locale.get(),
                 ImmutableMap.copyOf(sessionProperties),
@@ -603,59 +591,6 @@ public class PrestoConnection
         if (isClosed()) {
             throw new SQLException("Connection is closed");
         }
-    }
-
-    private void setCatalogAndSchema()
-            throws SQLException
-    {
-        String path = uri.getPath();
-        if (path.equals("/")) {
-            return;
-        }
-
-        // remove first slash
-        if (!path.startsWith("/")) {
-            throw new SQLException("Path does not start with a slash: " + uri);
-        }
-        path = path.substring(1);
-
-        List<String> parts = Splitter.on("/").splitToList(path);
-
-        // remove last item due to a trailing slash
-        if (parts.get(parts.size() - 1).isEmpty()) {
-            parts = parts.subList(0, parts.size() - 1);
-        }
-
-        if (parts.size() > 2) {
-            throw new SQLException("Invalid path segments in URL: " + uri);
-        }
-
-        if (parts.get(0).isEmpty()) {
-            throw new SQLException("Catalog name is empty: " + uri);
-        }
-        catalog.set(parts.get(0));
-
-        if (parts.size() > 1) {
-            if (parts.get(1).isEmpty()) {
-                throw new SQLException("Schema name is empty: " + uri);
-            }
-            schema.set(parts.get(1));
-        }
-    }
-
-    @VisibleForTesting
-    URI getHttpUri()
-    {
-        return createHttpUri(address);
-    }
-
-    private static URI createHttpUri(HostAndPort address)
-    {
-        return uriBuilder()
-                .scheme((address.getPort() == 443) ? "https" : "http")
-                .host(address.getHostText())
-                .port(address.getPort())
-                .build();
     }
 
     private static void checkResultSet(int resultSetType, int resultSetConcurrency)
