+++ presto-jdbc/src/test/java/com/facebook/presto/jdbc/TestProgressMonitor.java
@@ -119,7 +119,7 @@ public class TestProgressMonitor
         HttpClient client = new TestingHttpClient(new TestingHttpClientProcessor(RESPONSES));
         QueryExecutor testQueryExecutor = QueryExecutor.create(client);
         URI uri = URI.create(format("prestotest://%s", SERVER_ADDRESS));
-        return new PrestoConnection(uri, "test", testQueryExecutor);
+        return new PrestoConnection(new ConnectionParameters(uri), "test", testQueryExecutor);
     }
 
     private static class TestingHttpClientProcessor
